Move tests into test_cmodule.py
